+++ src/main/java/com/morningstar/Calculators/CalcApplication.java
@@ -15,8 +15,8 @@ public class CalcApplication implements CalcInterface {
 		return (input1 * input2);
 	}
 
-	public Double divide(double input1, double input2) {
-		if (input2 != 0) {
+	public Double divide(float input1, float input2) {
+		if (input2 != 0.0f) {
 			double x = (input1 / input2);
 			return x;
 		} else {
